+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+<br>
+
 ## Unreleased
 * `QueryError`: `messages` was renamed to `remarks`
 * Add `QueryResponseError`, which is raised for unexpected query responses
@@ -10,6 +12,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   its subclasses are raised
 * The `repr()` of errors have slightly changed
 * Fix incorrect documentation of `collect_routes()`
+* Add `concurrency` parameter to `Client`: you can now specify a concurrent
+  query limit that caps the server's own, or replaces it if there is none.
+* Add `concurrency` field to `Status`
+
+<br>
 
 ## [0.1.2] - 2023-04-27
 * Fix zero timeout which would previously be interpreted as "no timeout"
@@ -18,6 +25,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   aio_overpass.error.QueryLanguageError: query <no kwargs> failed: 'line 1: parse error: Key expected - '%' found.' (+14 more)
   ```
 * [0.1.2.post1]: releases are now automatically published to PyPI when pushing a new tag
+
+<br>
 
 ## [0.1.1] - 2023-04-25
 * Fix wrong coordinate order in elements' GeoJSON `bbox`